dev 2/architercture/lab1/backend/requirements.txt
djangorestframework
markdown
django-filter
psycopg[binary,pool]
django-extensions
python-dotenv
gunicorn